move test to main pr test
